requirements.txt
asgiref==3.5.2
autopep8==1.7.0
Django==4.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
flake8==5.0.4
mccabe==0.7.0
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.4.0
python-dotenv==0.20.0
pytz==2022.2.1
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.2
